# static/swagger.yaml
openapi: 3.0.1
info:
  title: Application
  description: Generated at 24/09/2023 20:29:27. This is the swagger 
    ui based on the open api 3.0 specification of the Application
  version: 1.0.0 created by the flask swagger generator.
...
servers:
  - url: '/'
paths:
...
  '/student/{item_id}':
    get:
      tags:
      - None
      operationId: 'all_students'
      parameters:
      - in: path
        name: item_id
        schema:
          type: integer
        description: None
        required: True
      responses:
        '200':
          $ref: '#/components/responses/all_students_response'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    all_students_response:
      description: all_students response
      content:
         application/json:
             schema:
                 $ref: '#/components/schemas/all_students_response_schema'
  schemas:
...
    all_students_response_schema:
      type: object
      properties:
        age:
          type: integer  
          example: 32  
        bio:
          type: string  
          example: INITIAL BIO  
        created_at:
          type: string  
          example: 2023-08-01 18:47:02  
        email:
          type: string  
          example: s@test.com  
        firstname:
          type: string  
          example: Sudarshan  
        id:
          type: integer  
          example: 1  
        lastname:
          type: string  
          example: Shrestha  